Adding changes

+++ src/controllers/dashboardControllers/deleteExpenseType.js
@@ -1,17 +1,35 @@
 const ExpenseType = require('../../models/expenseType')
 const response = require('../../utils/response')
+const Expense=require("../../models/expense")
 
 const deleteExpenseType = async (req, res) => {
     const {expenseTypeId} = req.body 
     const { _id } = req.userDetails
     try{
-       const deleteExpenseType=await ExpenseType.findOneAndRemove({_id:expenseTypeId});
-       if(deleteExpenseType)
+       const findExpenseType=await ExpenseType.findOne({_id:expenseTypeId});
+       if(findExpenseType)
        {
-        return response(res, 200, true, "Expense Type deleted Successfully",null);
+        const deleteExpense=await Expense.deleteMany({expenseType:findExpenseType._id})
+        if(deleteExpense)
+        {
+          const deleteExpenseType=await ExpenseType.findOneAndRemove({_id:expenseTypeId})
+          if(deleteExpenseType)
+          {
+            return response(res, 200, true, "Expense Type deleted Successfully",{deleteExpenseType});
+          }
+          else
+          {
+            return response(res, 200, false, "Expense Type cannot deleted",null)
+          }
+          
+        }
+        else{
+          return response(res, 200, false, "Expense Type cannot deleted",null);  
+        }
+        
        }
        else{
-        return response(res, 200, false, "Expense Type cannot deleted",null);
+        return response(res, 200, false, "Expense Type not found",null);
        }
   }
   catch(error)
